requirements.txt
Flask==2.3.2
requests==2.31.0
cohere==4.11
python-whois==0.9.0
beautifulsoup4==4.12.2